Renaming repo to clijx-parallel
