+++ pom.xml
@@ -13,10 +13,10 @@
 	</parent>
 
 	<groupId>net.haesleinhuepf</groupId>
-	<artifactId>clijx-faclon-heavy_</artifactId>
+	<artifactId>clijx-parallel_</artifactId>
 	<version>0.30.1.14-SNAPSHOT</version>
 
-	<name>clijx-faclon-heavy_</name>
+	<name>clijx-parallel_</name>
 	<description>Multi-GPU support for processing big data</description>
 	<url>http://haesleinhuepf.net</url>
 	<inceptionYear>2021</inceptionYear>
