+++ src/test/java/net/haesleinhuepf/clijx/parallel/implementations/DummyFilter.java
@@ -1,7 +1,7 @@
-package net.haesleinhuepf.clijx.faclonheavy.implementations;
+package net.haesleinhuepf.clijx.parallel.implementations;
 
 import net.haesleinhuepf.clij.clearcl.ClearCLBuffer;
-import net.haesleinhuepf.clijx.faclonheavy.AbstractTileProcessor;
+import net.haesleinhuepf.clijx.parallel.AbstractTileProcessor;
 
 import java.util.Arrays;
